Refactored tests to run in isolation
